refactor out select object
